refactor: extract comon function to test support file
