+++ repository/support_test.go
@@ -1,0 +1,19 @@
+package repository
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+	"github.com/juliocnsouzadev/kafka-ish/model"
+)
+
+func getMessage() model.Message {
+	id := uuid.New().String()
+	message := model.Message{
+		Id:        id,
+		Topic:     "topic",
+		Timestamp: time.Now(),
+		Content:   "content",
+	}
+	return message
+}
